# Improve local search performance