@niheaven 2022-03-17
Sorry, but this approach took longer time in my PC:

![image](https://user-images.githubusercontent.com/5832170/158753254-408a4b11-a5af-4b97-b272-caa1d1d7caf2.png)

![image](https://user-images.githubusercontent.com/5832170/158753293-cf7433af-2db3-47d2-bc9a-f8f4e8507631.png)

And I've tried several times. What about yours @rashil2000 ?