@bartvanandel 2022-03-17
Strange. On my system, `scoop search net` went down from 32 seconds to 4 seconds with this change. Likewise, `scoop search zulu` went from 8 seconds to 2 seconds (this contains hits in the java bucket).

BTW due to (presumably) file system caching, next runs of the same command and on the same develop branch are significantly faster. So measuring this in a fair way may be tricky, especially since in practice, usually you don't search the same thing twice in a row.