@rashil2000 2022-03-18
> Search twice or use different methods?

Different methods. See `.\bin\scoop search zulu` vs `scoop search zulu`