@chawyehsu 2022-05-16
```
DESKTOP in current on  develop
❯ scoop bucket list

Name     Source                                     Updated            Manifests
----     ------                                     -------            ---------
akira    https://github.com/chawyehsu/scoop-akira   2022/4/13 23:52:50         2
dorado   https://github.com/chawyehsu/dorado        2022/5/13 17:08:19       217
extras   https://github.com/ScoopInstaller/Extras   2022/5/16 12:30:48      1501
java     https://github.com/ScoopInstaller/Java     2022/5/15 4:01:48        220
main     https://github.com/ScoopInstaller/Main     2022/5/16 0:30:16       1019
versions https://github.com/ScoopInstaller/Versions 2022/5/16 15:36:44       314

DESKTOP in current on  develop
❯ Measure-Command { .\bin\scoop.ps1 search rstudio }
'extras' bucket:
'versions' bucket:

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 6
Milliseconds      : 911
Ticks             : 69116116
TotalDays         : 7.99955046296296E-05
TotalHours        : 0.00191989211111111
TotalMinutes      : 0.115193526666667
TotalSeconds      : 6.9116116
TotalMilliseconds : 6911.6116


DESKTOP in current on  develop took 6s
❯ git sw fix/4239_searchPerformance
Switched to branch 'fix/4239_searchPerformance'
DESKTOP in current on  fix/4239_searchPerformance
❯ Measure-Command { .\bin\scoop.ps1 search rstudio }
'extras' bucket:
'versions' bucket:

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 1
Milliseconds      : 320
Ticks             : 13204317
TotalDays         : 1.52827743055556E-05
TotalHours        : 0.000366786583333333
TotalMinutes      : 0.022007195
TotalSeconds      : 1.3204317
TotalMilliseconds : 1320.4317


```