@niheaven 2022-05-17
Test it again, same result:

```console
Scoop\apps\scoop took 23s
❯ git -C "current" checkout develop
Switched to branch 'develop'
Your branch is behind 'origin/develop' by 1 commit, and can be fast-forwarded.
  (use "git pull" to update your local branch)

Scoop\apps\scoop
❯ Measure-Command { scoop search rstudio }
'extras' bucket:
'versions' bucket:

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 10
Milliseconds      : 352
Ticks             : 103529888
TotalDays         : 0.000119826259259259
TotalHours        : 0.00287583022222222
TotalMinutes      : 0.172549813333333
TotalSeconds      : 10.3529888
TotalMilliseconds : 10352.9888



Scoop\apps\scoop took 10s
❯ git -C "current" checkout fix/4239_searchPerformance
Switched to branch 'fix/4239_searchPerformance'

Scoop\apps\scoop
❯ Measure-Command { scoop search rstudio }
'extras' bucket:
'versions' bucket:

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 23
Milliseconds      : 881
Ticks             : 238818129
TotalDays         : 0.000276409871527778
TotalHours        : 0.00663383691666667
TotalMinutes      : 0.398030215
TotalSeconds      : 23.8818129
TotalMilliseconds : 23881.8129
```

Really strange...